requirements.txt
# Core Python utilities
fastapi
uvicorn
groq
python-dotenv
groq
pathlib
typing-extensions>=4.5.0

# Text Extraction
pdfplumber==0.11.0
python-docx==1.1.0

# Data handling
jsonschema>=4.19.0

# Logging & utilities
tqdm>=4.66.0

# Evaluation (rule-based extraction metrics)
scikit-learn>=1.3.0

# LLM client for Azure OpenAI-compatible endpoint
openai>=1.40.0

# Environment variables (.env)
python-dotenv>=1.0.0

# HTTP requests (if you use requests elsewhere)
requests>=2.25.0

# --- Optional API / Deployment (keep commented if not used) ---
# fastapi>=0.103.0
# uvicorn>=0.23.2
# pydantic>=2.4.0
